Option tag to identify efs by tag
the option 'tag' allow to identify an efs file system from the tag. This option, allow to have a more friendly /etc/fstab, and in case or inter region disaster recovery, you can have 2 efs with the same tag, /etc/fstab can be the same in both region. In any case, the retrieval of the file-system-id is done only in the actual region.